refact: 훅 합치고 변수명, 함수명 수정
